refactor: split assets.py into assets/ package (model, manager, files, labels mixins)
